added hash guards to workflow execution (i.e. on workflow inputs)
